app/requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
Django==3.0.7
djangorestframework==3.12.2
djangorestframework-simplejwt
idna==2.10
PyJWT==1.7.1
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
django-heroku
gunicorn==20.0.4
whitenoise